fix(task-archive): drop the cross-module import — these files are loaded BY PATH
Found the mechanism behind #2819's CI failure, and it is mine. `local_task_protocol`
imported `newest_archived` from `task_archive`. Several consumers load BOTH modules
with `spec_from_file_location` and `src/` NOT on `sys.path`, and there any import of
the sibling raises ModuleNotFoundError and takes its caller down with it. In the
notifier path that kills the workstream-context build, whose stderr is discarded
(`task-notifier.sh:83` redirects to /dev/null), so the capture file simply never
appears — the FileNotFoundError CI reports.

Reproduced directly rather than by inference: after I inverted the import to
`task_archive -> local_task_protocol`, `discord-bridge-archive-failure-keeps-the-file`
failed with exactly that ModuleNotFoundError, at the line the new import added. Same
defect, opposite direction, on this machine.

So: no cross-import in either direction. Collision NAMING stays in task_archive
(`_move_without_clobbering` mints `.N`); collision SELECTION moves to
local_task_protocol, where `find_archived_task` — its only caller — already lives.
One owner each, and neither module needs the other in order to load. This also
removes the sys.path mutation the lazy import performed on first call.

Controls:
  no-scan guard   FAILS against a glob impl, passes here (still validated)
  import shapes   flat AND package both resolve; sys.path unchanged
  regression      discord-bridge-archive-failure-keeps-the-file back to OK
  task-archive    17/17; ag2-sparrow mirror re-synced; resolution gate clean

Whether this is the WHOLE cause is CI's to say — the failing test cannot run on this
host at all (it errors at origin/main alone here), so CI is the only instrument.

Stand: PRO
